Move get_blockstore_type to utils
