+++ src/lakefs_spec/util.py
@@ -1,6 +1,7 @@
 import re
 
 from lakefs_client import __version__ as __lakefs_client_version__
+from lakefs_client.client import LakeFSClient
 
 
 def parse(path: str) -> tuple[str, str, str]:
@@ -36,3 +37,21 @@ def parse(path: str) -> tuple[str, str, str]:
 
 lakefs_client_version = tuple(int(v) for v in __lakefs_client_version__.split("."))
 del __lakefs_client_version__
+
+
+def get_blockstore_type(client: LakeFSClient) -> str:
+    """
+    Get the blockstore of the lakeFS server.
+    Backwards compatible to breaking config_api change in lakeFSClient version 0.110.1.
+
+    Args:
+        client (LakeFSClient): The lakefs client.
+
+    Returns:
+        str: The lakeFS server's blockstore type.
+    """
+    if lakefs_client_version < (0, 111, 0):
+        blockstore_type = client.config_api.get_storage_config().blockstore_type
+    else:
+        blockstore_type = client.config_api.get_config().storage_config.blockstore_type
+    return blockstore_type
